Commit source for most recent post.
After the disasterous commit last week that inadvertantly
destroyed my blog site (more specifically, destroyed the
host-name resolution via blog.pnkfx.org), here is the source
code associated with that post.

I suppose I could actually make a blog post about the mistake I made
tha caused the destruction.  The easy way to sum it up is: The model
employed by octopress when deploying to github is this: Your _deploy/
subdirectory must contain a checkout of the target repo, the one with
the gh-pages branch, and you must have that _deploy/ subdirectory
checked out and ready to go before running 'rake deploy'.

If you do not have a _deploy/ subdirectory at all and you let 'rake
deploy' create it for you but you also let 'rake deploy' attempt to
push to github, and you are also managing the source itself on github,
you will enter a world of pain where the rake invocation will push
*this* root directory, presumably in the master branch (or in my case,
blog.pnkfx.org branch) to the target repo in the gh-pages/ branch.
Which will bust things terribly, especially if that causes the CNAME
file to get deleted from the gh-pages/ branch of the target repo.