Fixed `lookahead-matches?` bug.
Added origin-tracking information to states so I can tell why they
were added.  The field is generalizable to arbitrary S-exp's, and
really the intention is that I could plug-in the tree to justify it;
such back-pointers might also be useful for building up *a* parse-tree
(but it might not be as helpful for building *all* parse trees).

Added instrumentation tracking points when states are added to the
sets, as well as when completor considers doing such additions.